+++ recipes/default.rb
@@ -1,10 +1,10 @@
 #
-# Cookbook Name:: syncdata-nativex
+# Cookbook Name:: syncdata-simplyadrian
 # Recipe:: default
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
 
-include_recipe "syncdata-nativex::ship_ssh_keys"
+include_recipe "syncdata-simplyadrian::ship_ssh_keys"