+++ recipes/pull-datasync.rb
@@ -1,8 +1,8 @@
 #
-# Cookbook Name:: syncdata-nativex
+# Cookbook Name:: syncdata-simplyadrian
 # Recipe:: pull-datasync
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #