+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-syncdata-nativex CHANGELOG
+syncdata-simplyadrian CHANGELOG
 ==========================
 
-This file is used to list changes made in each version of the syncdata-nativex cookbook.
+This file is used to list changes made in each version of the syncdata-simplyadrian cookbook.
 
 0.1.0
 -----
-- Adrian Herrera - Initial release of syncdata-nativex
+- Adrian Herrera - Initial release of syncdata-simplyadrian
 
 0.1.1
 -----